refactor(front): :recycle: Refactor EventList with Pinia store and API service
